+++ EasyPost/Exceptions/General/MissingPropertyError.cs
@@ -1,4 +1,3 @@
-using System.Diagnostics.CodeAnalysis;
 using System.Globalization;
 
 namespace EasyPost.Exceptions.General