+++ EasyPost/Parameters/CarrierAccount/Create.cs
@@ -7,7 +7,7 @@ using EasyPost.Utilities.Internal.Attributes;
 namespace EasyPost.Parameters.CarrierAccount
 {
     /// <summary>
-    ///     <a href="https://www.easypost.com/docs/api#create-a-carrier-account">Parameters</a> for <see cref="EasyPost.Services.CarrierAccountService.Create(Create, System.Threading.CancellationToken)"/> API calls.
+    ///     <a href="https://www.easypost.com/docs/api#create-a-carrier-account">Parameters</a> for <see cref="EasyPost.Services.CarrierAccountService.Create(ACreate, System.Threading.CancellationToken)"/> API calls.
     /// </summary>
     [ExcludeFromCodeCoverage]
     public class Create : ACreate