+++ EasyPost/Parameters/Webhook/Update.cs
@@ -14,7 +14,7 @@ namespace EasyPost.Parameters.Webhook
 
         /// <summary>
         ///     The new secret value that will be used to generate a HMAC-SHA256 signature, included in the headers for each webhook payload event.
-        ///     Use this to verify an incoming webhook is from EasyPost, via <see cref="Services.WebhookService.ValidateWebhook(byte[], Dictionary{string,object}, string)"/>
+        ///     Use this to verify an incoming webhook is from EasyPost, via <see cref="Services.WebhookService.ValidateWebhook(byte[], Dictionary{string,object}, string)"/>.
         /// </summary>
         [TopLevelRequestParameter(Necessity.Optional, "webhook_secret")]
         public string? Secret { get; set; }